Adds ObservableFactory to auto-generate Observables
